dedicated Copilot default ProgramPatcher
